requirements.txt
anyio==4.0.0
asgiref==3.7.2
base58==2.1.1
certifi==2023.7.22
crispy-bootstrap4==2022.1
Deprecated==1.2.14
Django==4.2.6
django-crispy-forms==2.0
ECPy==1.2.5
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
gunicorn==20.1.0
pycryptodome==3.19.0
sniffio==1.3.0
sqlparse==0.4.4
types-Deprecated==1.2.9.3
typing_extensions==4.8.0
websockets==10.4
wrapt==1.15.0
xrpl-py==2.4.0
whitenoise==6.2.0
Pillow==9.3.0
